partial implements of login and logout and authentication in events routes

+++ server/models/eventsDAO.js
@@ -1,16 +1,21 @@
 const mongoose = require('mongoose')
 
-const {events} = require('../config/db')
+const {events, user} = require('../config/db')
+const {decodeJWT} = require('./auth/authUser')
 
 exports.get = (req, res, next) => {
+
+    const current_user = decodeJWT(req.headers['authorization'])
+
     events
-        .find()
-            .sort({dateOcurr: 1})
-        .then(data => {
-            res.status(200).json(data)
-        }).catch(err => {
-            res.status(400).json(err)
-        })
+        .find({whoCreated : current_user._id})
+        .sort({dateOcurr: -1})
+        .exec((err, result) => {
+            if(err) res.status(400).json(err)
+            else{
+                res.status(200).json({result})
+            }
+        })    
 }
 
 exports.getMonth = (req, res, next) => {
@@ -18,8 +23,10 @@ exports.getMonth = (req, res, next) => {
     let month = req.params.month -1;
     let year = req.params.year;
 
+    const current_user = decodeJWT(req.headers['authorization'])
+
     events
-        .find()
+        .find({whoCreated : current_user._id})
             .sort({dateOcurr: 1})
         .then(data => {
             let filter = data.filter( event => event.dateOcurr.getUTCMonth() == month && event.dateOcurr.getFullYear() == year);
@@ -33,19 +40,37 @@ exports.post = (req, res, next) => {
     let newEvent = new events(req.body)
     console.log(req.body)
 
+    const current_user = decodeJWT(req.headers['authorization'])
+
+    newEvent = current_user._id
+
     newEvent
         .save()
-        .then(e => {
-            res.status(200).json('Successfuly request')
-        }).catch(err => {
-            console.log(err)
-            res.status(400).json('Internal server error')
-        })
+        .then(response =>
+            user.findOne({_id: current_user._id}, (err, result) => {
+                result.events.push(response._id)
+                result.save()
+                    .then(response =>  res.status(200).json('successfuly request'))
+                    .catch(err => res.status(500).json('Internal server error'))
+            }) 
+           )
+}
+
+exports.update = (req,res, next) => {
+    events.updateOne(
+        {_id:mongoose.Types.ObjectId(req.body._id) },
+        {
+            ... req.body
+        },
+        (err, response) => err ? res.status(500).json(error) : res.status(200).json(response))
 }
 
 exports.delete = (req, res, next) => {
+
+    const current_user = decodeJWT(req.headers['authorization'])
+
     events
-        .deleteOne({'_id': req.params.id})
+        .deleteOne({whoCreated: current_user._id , _id: req.params.id})
         .then(data => {
             res.status(200).json(data)
         }).catch(err => {
